update set_hook location
